Split pendula robot from the pendula envs due to changes in the underlying mujoco xml base env.
